+++ androidx/src/main/java/me/panpf/androidx/util/NullableResultRunnable.java
@@ -1,0 +1,8 @@
+package me.panpf.androidx.util;
+
+import android.support.annotation.Nullable;
+
+public interface NullableResultRunnable<T> {
+    @Nullable
+    T run();
+}