+++ androidx/src/main/java/me/panpf/androidx/util/ResultRunnable.java
@@ -1,0 +1,8 @@
+package me.panpf.androidx.util;
+
+import android.support.annotation.NonNull;
+
+public interface ResultRunnable<T> {
+    @NonNull
+    T run();
+}